+++ src/vue/fields/AmountInputField.vue
@@ -47,6 +47,7 @@ import { vuexTypes } from '@/vuex'
 import config from '@/config'
 import { required, amountRange } from '@validators'
 import { inputStepByDigitsCount } from '@/js/helpers/input-trailing-digits-count'
+import { ErrorHandler } from '@/js/helpers/error-handler'
 
 const EVENTS = {
   input: 'input',
@@ -73,10 +74,9 @@ export default {
     value: { type: [Number, String], default: undefined },
     validationType: { type: String, default: '' },
     isMaxButtonShown: { type: Boolean, default: false },
-    min: { type: [Number, String], default: config.MIN_AMOUNT },
+    min: { type: [Number, String], required: true },
     max: { type: [Number, String], default: config.MAX_AMOUNT },
   },
-
   data: _ => ({
     EVENTS,
   }),
@@ -89,7 +89,6 @@ export default {
       },
     }
   },
-
   computed: {
     ...mapGetters([
       vuexTypes.assetByCode,
@@ -109,8 +108,8 @@ export default {
     },
 
     step () {
-      return this.assetRecord.trailingDigits
-        ? inputStepByDigitsCount(this.assetRecord.trailingDigits)
+      return this.assetRecord.trailingDigitsCount
+        ? inputStepByDigitsCount(this.assetRecord.trailingDigitsCount)
         : inputStepByDigitsCount(config.DECIMAL_POINTS)
     },
 
@@ -164,6 +163,12 @@ export default {
         },
       }
     },
+  },
+  created () {
+    if (this.min && this.min < this.step) {
+      ErrorHandler.processWithoutFeedback(
+        new Error('Min value cannot be less than step'))
+    }
   },
 }
 </script>